Simplifying check logic
I was trying to simplify the logic of the function. Is it really necessary to check for 'package' and overwrite everything in the file if the statement isn't found? Shouldn't it be easier—something like: file is empty, write the package statement; file has something inside, leave it like it is. Am i missing something?